NN-Dateien verschoben, nicht benötigte Dateien gelöscht
