+++ Kalaha/Kalaha.csproj
@@ -91,8 +91,6 @@
       <DependentUpon>KalahaBoardDisplaySTD.cs</DependentUpon>
     </EmbeddedResource>
   </ItemGroup>
-  <ItemGroup>
-    <Folder Include="NeuralNetwork" />
-  </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
