+++ Kalaha/NeuralNetwork/SinglePerceptron.cs
@@ -1,0 +1,31 @@
+﻿/*
+ * Created by SharpDevelop.
+ * User: Daniel
+ * Date: 23.09.2017
+ * Time: 00:38
+ * 
+ * To change this template use Tools | Options | Coding | Edit Standard Headers.
+ */
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+//-------UNDER CONSTRUCTION-----------
+
+//Evt noch "papaklasse" abstract particularNN
+
+namespace NeuralNetwork
+{
+	/// <summary>
+	/// Description of SinglePerceptron.
+	/// </summary>
+	public class SinglePerceptron
+	{
+//		private List<InputNeuron> inputNeurons = new List<InputNeuron>();
+//		private List<WorkingNeuron> workingNeurons = new List<WorkingNeuron>();
+		
+		//public NeuralNetwork NN = new NeuralNetwork();
+	}
+}
