--- NeuralNetwork/NeuralNetwork.sln
@@ -1,18 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 11.00
-# Visual Studio 2010
-# SharpDevelop 5.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NeuralNetwork", "NeuralNetwork.csproj", "{F45E14E0-554F-46F4-9347-47420434AC36}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{F45E14E0-554F-46F4-9347-47420434AC36}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F45E14E0-554F-46F4-9347-47420434AC36}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F45E14E0-554F-46F4-9347-47420434AC36}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F45E14E0-554F-46F4-9347-47420434AC36}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-EndGlobal
